+++ src/VisualStudio/Core/Def/Implementation/LanguageService/AbstractPackage`2.cs
@@ -75,6 +75,8 @@ namespace Microsoft.VisualStudio.LanguageServices.Implementation.LanguageService
 
                 // start remote host
                 EnableRemoteHostClientService();
+
+                Workspace.AdviseSolutionEvents(GetService(typeof(SVsSolution)) as IVsSolution);
             }
 
             // Ensure services that must be created on the UI thread have been.
